+++ CHANGELOG.md
@@ -37,6 +37,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 ### Fixed
   - Fix nginx gateway timeout on unshelve
     ([#686](https://github.com/cyverse/atmosphere/pull/686))
+  - Fix exception handling preventing instance destroy
+    ([#690](https://github.com/cyverse/atmosphere/pull/690))
+
 
 ## [v34-4](https://github.com/cyverse/atmosphere/compare/v34-3...v34-4) - 2018-09-25
 ### Fixed